+++ vlib/time/time_test.v
@@ -143,7 +143,7 @@ fn test_add_days() {
 	num_of_days := 3
 	t := time_to_test.add_days(num_of_days)
 	assert t.day == time_to_test.day + num_of_days
-	assert t.unix == time_to_test.unix + 86400 * num_of_days
+	assert t.unix == time_to_test.unix + 86400 * u64(num_of_days)
 }
 
 fn test_str() {
